Cleaned up scene and naming convention. #4
