+++ Camera-2D-Project/Assets/Code/FQ.Camera.meta
