+++ Camera-2D-Project/Assets/Code/FQ.Camera/FollowCamera/API/FollowCameraAPI.asmdef
@@ -1,6 +1,6 @@
 {
     "name": "FollowCameraAPI",
-    "rootNamespace": "FollowCamera",
+    "rootNamespace": "",
     "references": [],
     "includePlatforms": [],
     "excludePlatforms": [],
