+++ Camera-2D-Project/Assets/Code/FQ.Camera/FollowCamera/API/IFollowCamera.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Code.FQCamera.FollowCamera
+namespace Code.FQ.Camera.FollowCamera
 {
     /// <summary>
     /// A game object with the aim to follow another camera object around.
