+++ Camera-2D-Project/Assets/Code/FQ.UnityAPI/GameStatics/API/IUnityTime.cs
