+++ Camera-2D-Project/Assets/Code/FQ.UnityAPI/GameStatics/UnityTime.cs
